start

+++ handlers/callback_queries/client.go
@@ -2,6 +2,7 @@ package callback_queries
 
 import (
 	"context"
+	"fmt"
 	tg_bot "github.com/go-telegram/bot"
 	tg_models "github.com/go-telegram/bot/models"
 	"go.uber.org/dig"
@@ -13,7 +14,9 @@ import (
 	"rezvin-pro-bot/services"
 	bot_utils "rezvin-pro-bot/utils/bot"
 	utils_context "rezvin-pro-bot/utils/context"
+	"rezvin-pro-bot/utils/inline_keyboards"
 	"rezvin-pro-bot/utils/messages"
+	validate_data "rezvin-pro-bot/utils/validate"
 	"strings"
 )
 
@@ -25,7 +28,7 @@ type clientHandlerDependencies struct {
 	dig.In
 
 	Logger                       logger.ILogger                             `name:"Logger"`
-	InlineKeyboardService        services.IInlineKeyboardService            `name:"InlineKeyboardService"`
+	ConversationService          services.IConversationService              `name:"ConversationService"`
 	UserRepository               repositories.IUserRepository               `name:"UserRepository"`
 	ProgramRepository            repositories.IProgramRepository            `name:"ProgramRepository"`
 	UserProgramRepository        repositories.IUserProgramRepository        `name:"UserProgramRepository"`
@@ -34,7 +37,7 @@ type clientHandlerDependencies struct {
 
 type clientHandler struct {
 	logger                       logger.ILogger
-	inlineKeyboardService        services.IInlineKeyboardService
+	conversationService          services.IConversationService
 	userRepository               repositories.IUserRepository
 	programRepository            repositories.IProgramRepository
 	userProgramRepository        repositories.IUserProgramRepository
@@ -44,7 +47,7 @@ type clientHandler struct {
 func NewClientHandler(deps clientHandlerDependencies) *clientHandler {
 	return &clientHandler{
 		logger:                       deps.Logger,
-		inlineKeyboardService:        deps.InlineKeyboardService,
+		conversationService:          deps.ConversationService,
 		userRepository:               deps.UserRepository,
 		programRepository:            deps.ProgramRepository,
 		userProgramRepository:        deps.UserProgramRepository,
@@ -90,9 +93,19 @@ func (h *clientHandler) Handle(ctx context.Context, b *tg_bot.Bot, update *tg_mo
 		return
 	}
 
-	switch callBackQueryData {
-	case callback_data.ClientList:
+	if strings.HasPrefix(callBackQueryData, callback_data.ClientResultModifyList) {
+		h.resultModifyList(ctx, b)
+		return
+	}
+
+	if strings.HasPrefix(callBackQueryData, callback_data.ClientResultModifySelected) {
+		h.resultModifySelected(ctx, b)
+		return
+	}
+
+	if strings.HasPrefix(callBackQueryData, callback_data.ClientList) {
 		h.list(ctx, b)
+		return
 	}
 }
 
@@ -108,7 +121,11 @@ func (h *clientHandler) list(ctx context.Context, b *tg_bot.Bot) {
 		return
 	}
 
-	bot_utils.SendMessageWithInlineKeyboard(ctx, b, chatId, messages.SelectClientMessage(), h.inlineKeyboardService.ClientList(clients))
+	clientCount := h.userRepository.CountClients(ctx)
+
+	kb := inline_keyboards.ClientList(clients, clientCount, limit, offset)
+
+	bot_utils.SendMessageWithInlineKeyboard(ctx, b, chatId, messages.SelectClientMessage(), kb)
 }
 
 func (h *clientHandler) selected(ctx context.Context, b *tg_bot.Bot) {
@@ -117,7 +134,7 @@ func (h *clientHandler) selected(ctx context.Context, b *tg_bot.Bot) {
 
 	msg := messages.SelectClientOptionMessage(user.GetPrivateName())
 
-	bot_utils.SendMessageWithInlineKeyboard(ctx, b, chatId, msg, h.inlineKeyboardService.ClientSelectedMenu(user.Id))
+	bot_utils.SendMessageWithInlineKeyboard(ctx, b, chatId, msg, inline_keyboards.ClientSelectedMenu(user.Id))
 }
 
 func (h *clientHandler) programSelected(ctx context.Context, b *tg_bot.Bot) {
@@ -125,8 +142,15 @@ func (h *clientHandler) programSelected(ctx context.Context, b *tg_bot.Bot) {
 	user := utils_context.GetUserFromContext(ctx)
 	userProgram := utils_context.GetUserProgramFromContext(ctx)
 
+	if userProgram.UserId != user.Id {
+		h.logger.Error(fmt.Sprintf("UserProgram %d not assigned for user %d", userProgram.Id, user.Id))
+		msg := messages.ClientProgramNotAssignedMessage(user.GetPrivateName(), userProgram.Name())
+		bot_utils.SendMessage(ctx, b, chatId, msg)
+		return
+	}
+
 	msg := messages.SelectClientProgramOptionMessage(user.GetPrivateName(), userProgram.Name())
-	kb := h.inlineKeyboardService.ClientSelectedProgramMenu(user.Id, *userProgram)
+	kb := inline_keyboards.ClientSelectedProgramMenu(user.Id, *userProgram)
 
 	bot_utils.SendMessageWithInlineKeyboard(ctx, b, chatId, msg, kb)
 }
@@ -135,6 +159,13 @@ func (h *clientHandler) programDelete(ctx context.Context, b *tg_bot.Bot) {
 	chatId := utils_context.GetChatIdFromContext(ctx)
 	user := utils_context.GetUserFromContext(ctx)
 	userProgram := utils_context.GetUserProgramFromContext(ctx)
+
+	if userProgram.UserId != user.Id {
+		h.logger.Error(fmt.Sprintf("UserProgram %d not assigned for user %d", userProgram.Id, user.Id))
+		msg := messages.ClientProgramNotAssignedMessage(user.GetPrivateName(), userProgram.Name())
+		bot_utils.SendMessage(ctx, b, chatId, msg)
+		return
+	}
 
 	h.userProgramRepository.DeleteById(ctx, userProgram.Id)
 	h.userExerciseRecordRepository.DeleteByUserProgramId(ctx, userProgram.Id)
@@ -159,9 +190,13 @@ func (h *clientHandler) programList(ctx context.Context, b *tg_bot.Bot) {
 		return
 	}
 
+	programsCount := h.userProgramRepository.CountAllByUserId(ctx, user.Id)
+
 	msg := messages.SelectClientProgramMessage(user.GetPrivateName())
 
-	bot_utils.SendMessageWithInlineKeyboard(ctx, b, chatId, msg, h.inlineKeyboardService.ClientProgramList(user.Id, programs))
+	kb := inline_keyboards.ClientProgramList(user.Id, programs, programsCount, limit, offset)
+
+	bot_utils.SendMessageWithInlineKeyboard(ctx, b, chatId, msg, kb)
 }
 
 func (h *clientHandler) programAdd(ctx context.Context, b *tg_bot.Bot) {
@@ -177,9 +212,13 @@ func (h *clientHandler) programAdd(ctx context.Context, b *tg_bot.Bot) {
 		return
 	}
 
+	programsCount := h.programRepository.CountNotAssignedToUser(ctx, user.Id)
+
 	msg := messages.SelectClientProgramMessage(user.GetPrivateName())
 
-	bot_utils.SendMessageWithInlineKeyboard(ctx, b, chatId, msg, h.inlineKeyboardService.ProgramForClientList(user.Id, programs))
+	kb := inline_keyboards.ProgramForClientList(user.Id, programs, programsCount, limit, offset)
+
+	bot_utils.SendMessageWithInlineKeyboard(ctx, b, chatId, msg, kb)
 }
 
 func (h *clientHandler) programAssign(ctx context.Context, b *tg_bot.Bot) {
@@ -222,7 +261,14 @@ func (h *clientHandler) resultList(ctx context.Context, b *tg_bot.Bot) {
 	user := utils_context.GetUserFromContext(ctx)
 	userProgram := utils_context.GetUserProgramFromContext(ctx)
 
-	records := h.userExerciseRecordRepository.GetByUserProgramId(ctx, userProgram.Id)
+	if userProgram.UserId != user.Id {
+		h.logger.Error(fmt.Sprintf("UserProgram %d not assigned for user %d", userProgram.Id, user.Id))
+		msg := messages.ClientProgramNotAssignedMessage(user.GetPrivateName(), userProgram.Name())
+		bot_utils.SendMessage(ctx, b, chatId, msg)
+		return
+	}
+
+	records := h.userExerciseRecordRepository.GetAllByUserProgramId(ctx, userProgram.Id)
 
 	if len(records) == 0 {
 		msg := messages.NoRecordsForClientProgramMessage(user.GetPrivateName(), userProgram.Name())
@@ -233,4 +279,75 @@ func (h *clientHandler) resultList(ctx context.Context, b *tg_bot.Bot) {
 	msg := messages.ClientProgramResultsMessage(user.GetPrivateName(), userProgram.Name(), records)
 
 	bot_utils.SendMessage(ctx, b, chatId, msg)
+
+	h.programSelected(ctx, b)
+}
+
+func (h *clientHandler) resultModifyList(ctx context.Context, b *tg_bot.Bot) {
+	chatId := utils_context.GetChatIdFromContext(ctx)
+	user := utils_context.GetUserFromContext(ctx)
+	userProgram := utils_context.GetUserProgramFromContext(ctx)
+	limit := utils_context.GetLimitFromContext(ctx)
+	offset := utils_context.GetOffsetFromContext(ctx)
+
+	if userProgram.UserId != user.Id {
+		h.logger.Error(fmt.Sprintf("UserProgram %d not assigned for user %d", userProgram.Id, user.Id))
+		msg := messages.ClientProgramNotAssignedMessage(user.GetPrivateName(), userProgram.Name())
+		bot_utils.SendMessage(ctx, b, chatId, msg)
+		return
+	}
+
+	records := h.userExerciseRecordRepository.GetByUserProgramId(ctx, userProgram.Id, limit, offset)
+
+	if len(records) == 0 {
+		msg := messages.NoRecordsForClientProgramMessage(user.GetPrivateName(), userProgram.Name())
+		bot_utils.SendMessage(ctx, b, chatId, msg)
+		return
+	}
+
+	recordsCount := h.userExerciseRecordRepository.CountAllByUserProgramId(ctx, userProgram.Id)
+
+	msg := messages.ClientProgramResultsModifyMessage(user.GetPrivateName(), userProgram.Name())
+
+	kb := inline_keyboards.ClientProgramResultsModifyList(user.Id, records, recordsCount, limit, offset)
+
+	bot_utils.SendMessageWithInlineKeyboard(ctx, b, chatId, msg, kb)
+}
+
+func (h *clientHandler) getWeight(ctx context.Context, b *tg_bot.Bot) int {
+	chatId := utils_context.GetChatIdFromContext(ctx)
+
+	conversation := h.conversationService.CreateConversation(chatId)
+	defer h.conversationService.DeleteConversation(chatId)
+
+	answer := conversation.WaitAnswer()
+
+	weight, err := validate_data.ValidateWeightAnswer(answer)
+
+	if err != nil {
+		bot_utils.SendMessage(ctx, b, chatId, err.Error())
+		return h.getWeight(ctx, b)
+	}
+
+	return weight
+}
+
+func (h *clientHandler) resultModifySelected(ctx context.Context, b *tg_bot.Bot) {
+	chatId := utils_context.GetChatIdFromContext(ctx)
+	user := utils_context.GetUserFromContext(ctx)
+	record := utils_context.GetUserExerciseRecordFromContext(ctx)
+
+	bot_utils.SendMessage(ctx, b, chatId, messages.EnterClientResultMessage(user.GetPrivateName(), record.Name()))
+
+	weight := h.getWeight(ctx, b)
+
+	h.userExerciseRecordRepository.UpdateById(ctx, record.Id, models.UserExerciseRecord{
+		Weight: weight,
+	})
+
+	msg := messages.ClientProgramResultModifiedMessage(user.GetPrivateName(), record.Name())
+
+	bot_utils.SendMessage(ctx, b, chatId, msg)
+
+	h.programSelected(ctx, b)
 }
